Refactor sync-cache command to read external repositories and MCP servers from manifest; update G4 client initialization in extension activation
